requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
SpeechRecognition==3.10.0
pydub==0.25.1
livekit==1.0.13
+